refactor(sales): split SalesOrderEdit (655→205) em hook + componentes
Extrai a lógica (load do pedido, edição de itens, addProduct, handleTintConfirm,
memos, handleSave com sync Omie fire-and-forget) para o hook useSalesOrderEdit e
quebra o JSX em PaymentComboboxEdit, AddProductSearch e OrderItemCard. Tipos em
types.ts. Comportamento 1:1; +12 testes.
